+++ src/helper.ts
@@ -1,21 +1,24 @@
 import { MarkdownView, App } from "obsidian";
-import { parse } from "path";
+import { parse } from "path-browserify";
 
 interface Image {
   path: string;
   name: string;
   source: string;
 }
-// ![](./dsa/aa.png) local image should has ext
+// ![](./dsa/aa.png) local image should has ext, support ![](<./dsa/aa.png>), support ![](image.png "alt")
 // ![](https://dasdasda) internet image should not has ext
-const REGEX_FILE = /\!\[(.*?)\]\((\S+\.\w+)\)|\!\[(.*?)\]\((https?:\/\/.*?)\)/g;
+const REGEX_FILE =
+  /\!\[(.*?)\]\(<(\S+\.\w+)>\)|\!\[(.*?)\]\((\S+\.\w+)(?:\s+"[^"]*")?\)|\!\[(.*?)\]\((https?:\/\/.*?)\)/g;
 const REGEX_WIKI_FILE = /\!\[\[(.*?)(\s*?\|.*?)?\]\]/g;
+
 export default class Helper {
   app: App;
 
   constructor(app: App) {
     this.app = app;
   }
+
   getFrontmatterValue(key: string, defaultValue: any = undefined) {
     const file = this.app.workspace.getActiveFile();
     if (!file) {
@@ -39,6 +42,7 @@ export default class Helper {
       return null;
     }
   }
+
   getValue() {
     const editor = this.getEditor();
     return editor.getValue();
@@ -60,6 +64,7 @@ export default class Helper {
     let value = editor.getValue();
     return this.getImageLink(value);
   }
+
   getImageLink(value: string): Image[] {
     const matches = value.matchAll(REGEX_FILE);
     const WikiMatches = value.matchAll(REGEX_WIKI_FILE);